+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.8.3 - February 2024
 
 - Add `sbp_member_id` and `sbp_receiver_pam` for payout method response
+- Add `is_test` for payout create method
 
 ## 1.8.2 - December 2023